MP: Moved test utility classes to main/tests package
