+++ shared/src/commonMain/kotlin/tests/README.md
@@ -1,0 +1,4 @@
+This package is needed for providing test utility classes in android/react modules. 
+If these packages reside in test instead of main then they cannot be used in other modules.
+
+[Reason](https://stackoverflow.com/questions/58956010/how-to-add-dependencies-to-tests-of-another-project-in-a-multi-platform-multi-pr)
